requirements.txt
CouchDB==0.8
Flask==0.9
FormEncode==1.2.6
Jinja2==2.6
MarkupSafe==0.15
Werkzeug==0.8.3
decorator==3.4.0
lxml==3.1.0
nose==1.2.1
pymediainfo==1.3.5
python-Levenshtein==0.10.2
python-memcached==1.48
simplejson==3.1.0
wsgiref==0.1.2
